requirements.txt
aiofiles==23.2.1
fastapi==0.104.0
Jinja2==3.1.2
openai==0.28.1
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn==0.23.2